+++ src/OpenTelemetry.Instrumentation.Runtime/MeterProviderBuilderExtensions.cs
@@ -1,8 +1,10 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+#if !NET9_0_OR_GREATER
 using OpenTelemetry.Instrumentation.Runtime;
 using OpenTelemetry.Internal;
+#endif
 
 namespace OpenTelemetry.Metrics;
 
@@ -11,6 +13,10 @@ namespace OpenTelemetry.Metrics;
 /// </summary>
 public static class MeterProviderBuilderExtensions
 {
+#if NET9_0_OR_GREATER
+    private const string DotNetRuntimeMeterName = "System.Runtime";
+#endif
+
     /// <summary>
     /// Enables runtime instrumentation.
     /// </summary>
@@ -18,8 +24,13 @@ public static class MeterProviderBuilderExtensions
     /// <returns>The instance of <see cref="MeterProviderBuilder"/> to chain the calls.</returns>
     public static MeterProviderBuilder AddRuntimeInstrumentation(
         this MeterProviderBuilder builder) =>
+#if NET9_0_OR_GREATER
+        builder.AddMeter(DotNetRuntimeMeterName);
+#else
         AddRuntimeInstrumentation(builder, configure: null);
+#endif
 
+#if !NET9_0_OR_GREATER
     /// <summary>
     /// Enables runtime instrumentation.
     /// </summary>
@@ -39,4 +50,5 @@ public static class MeterProviderBuilderExtensions
         builder.AddMeter(RuntimeMetrics.MeterInstance.Name);
         return builder.AddInstrumentation(() => instrumentation);
     }
+#endif
 }